Unified the rest of the projects

+++ engine/utils/memory/dlmalloc.c
@@ -672,7 +672,7 @@ static int init_mparams(void) {
 #elif defined(LACKS_TIME_H)
 			magic = (size_t)&magic ^ (size_t)0x55555555U;
 #else
-			magic = (size_t)(time(0) ^ (size_t)0x55555555U);
+			magic = (size_t)((size_t)time(0) ^ (size_t)0x55555555U);
 #endif
 			magic |= (size_t)8U;    /* ensure nonzero */
 			magic &= ~(size_t)7U;   /* improve chances of fault for bad values */
